Add data base H2, and fix pl message

+++ src/main/java/org/example/HelloService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public class HelloService {
     static final String FALLBACK_NAME = "world";
-    static final Lang FALLBACK_LANG = new Lang(1L, "Hello", "en");
+    static final Lang FALLBACK_LANG = new Lang(1, "Hello", "en");
     private final Logger logger = LoggerFactory.getLogger(HelloService.class);
 
     private LangRepository repository;
@@ -21,9 +21,9 @@ public class HelloService {
     }
 
     String prepareGreeting(String name, String lang) {
-        Long langId;
+        Integer langId;
         try {
-            langId = Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId());
+            langId = Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId());
         } catch (NumberFormatException e) {
             logger.warn("Non-numeric language id used: " + lang);
             langId = FALLBACK_LANG.getId();
